رفع برخی باگ و ...

+++ src/main/java/com/example/salaryproject/Employee.java
@@ -5,12 +5,12 @@ import java.time.LocalDate;
 public class Employee {
     private String firstName;
     private String lastName;
-    private int nationalId;
+    private long nationalId;
     private LocalDate dateOfBirth;
     private String email;
-    private int phoneNumber;
+    private long phoneNumber;
 
-    public Employee(String firstName, String lastName, int nationalId, LocalDate dateOfBirth, String email, int phoneNumber) {
+    public Employee(String firstName, String lastName, long nationalId, LocalDate dateOfBirth, String email, long phoneNumber) {
         this.firstName = firstName;
         this.lastName = lastName;
         this.nationalId = nationalId;
